third_party/WebKit/LayoutTests/fast/css/fontfaceset-add-remove-while-loading.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('Tests that FontFaceSet is correctly updated when added or removed a FontFace while loading.');

window.jsTestIsAsync = true;

function loadingDonePromise() {
    return new Promise(function(resolve) {
        function handler(e) {
            document.fonts.removeEventListener('loadingdone', handler);
            resolve(e);
        };
        document.fonts.addEventListener('loadingdone', handler);
    });
}

Promise.resolve().then(function() {
    face1 = new FontFace('face1', 'url(../../resources/Ahem.ttf)', {});
    shouldBeEqualToString('face1.status', 'unloaded');

    face1.load();
    shouldBeEqualToString('face1.status', 'loading');
    shouldBeEqualToString('document.fonts.status', 'loaded');

    document.fonts.add(face1);
    shouldBeEqualToString('document.fonts.status', 'loading');
    return document.fonts.ready;
}).then(function() {
    shouldBeEqualToString('face1.status', 'loaded');
    shouldBeEqualToString('document.fonts.status', 'loaded');

    face2 = new FontFace('face2', 'url(../../resources/Ahem.ttf)', {});
    shouldBeEqualToString('face2.status', 'unloaded');

    document.fonts.add(face2);
    var face2loaded = face2.load()
    shouldBeEqualToString('document.fonts.status', 'loading');

    var loadingdone = loadingDonePromise();
    document.fonts.delete(face2);
    shouldBeEqualToString('document.fonts.status', 'loading');
    shouldBeEqualToString('face2.status', 'loading');

    return Promise.all([face2loaded, loadingdone]);
}).then(function() {
    shouldBeEqualToString('face2.status', 'loaded');
    shouldBeEqualToString('document.fonts.status', 'loaded');

    face3 = new FontFace('face3', 'url(../../resources/Ahem.ttf)', {});
    shouldBeEqualToString('face3.status', 'unloaded');

    document.fonts.add(face3);
    face3.load();
    shouldBeEqualToString('document.fonts.status', 'loading');

    document.fonts.clear();
    shouldBeEqualToString('document.fonts.status', 'loading');
    shouldBeEqualToString('face3.status', 'loading');
    return document.fonts.ready;
}).then(function() {
    shouldBeEqualToString('document.fonts.status', 'loaded');
    finishJSTest();
}).catch(function (err) {
    testFailed('Unexpected rejection: ' + err);
    finishJSTest();
});

</script>
